feat: connect tradingview native chart controls
