+++ packages/kit/src/components/TradingView/TradingViewChartControls/index.ts
@@ -3,3 +3,9 @@ export {
   TradingViewChartControls,
 } from './TradingViewChartControls';
 export type { ITradingViewChartControlsProps } from './TradingViewChartControls';
+export {
+  IndicatorListDialogContent,
+  IndicatorPopover,
+} from './indicatorSelector/NativeIndicatorSelector';
+export { TradingViewNativeIntervalSelector } from './intervalSelector/NativeIntervalSelector';
+export type { ITradingViewNativeIntervalControlMode } from './intervalSelector/NativeIntervalSelector';
