+++ packages/kit/src/components/TradingView/TradingViewChartControls/intervalSelector/NativeIntervalSelector.tsx
@@ -20,7 +20,7 @@ import {
 } from './NativeIntervalUtils';
 
 import type { ITradingViewNativeIntervalControlMode } from './hooks/useNativeIntervalSelector';
-import type { ITradingViewIntervalConfigData } from '../../types';
+import type { ITradingViewIntervalConfigData } from '../../TradingViewV2/types';
 
 export type { ITradingViewNativeIntervalControlMode } from './hooks/useNativeIntervalSelector';
